Move StringSplitExtensions to its own file
Conditionally included only when TargetGroup != netcoreapp1.1.
